now also makes inverted index alongwith file comparison
opens lexicon,forward index and inverted index in append mode if a new file is added to the dataset. Keeps record of the files in the dataset by writing file names that are scanned to forward index, inverted index and lexicon to filecount.txt. Compares the dataset to the filecount.txt and if a new file is not added, simply exits the function. If filecount.txt is not present, opens everything in write mode and makes lexicon , filecount,inverted index and forward index. Still have to work on getting last added docid and word id.